+++ name/output.tf
@@ -1,9 +1,0 @@
-output "userDefinedString_replaced" {
-  description = "Returns the gccaf compliant name of resource group objects created"
-  value       = local.userDefinedString_replaced
-}
-
-output "names" {
-  description = "Returns the gccaf compliant name of resource group objects created"
-  value       = local.names
-}