+++ test/outputs.tf
@@ -1,5 +1,0 @@
-output "object" {
-  description = "Returns the full set of resource group objects created"
-
-  value = module.resource_groups_L1.1
-}